+++ src/main/java/com/banorte/ws/esb/reconciliaciones/service/ObtenerInventarioFullOutServiceImpl.java
@@ -26,7 +26,7 @@ public class ObtenerInventarioFullOutServiceImpl implements IObtenerInventarioFu
 	}
 
 	@Override
-	public ObtenerInventarioFullOut findByID(int id) {
+	public ObtenerInventarioFullOut findByID(Long id) {
 		return obtenerInventarioFullOutDao.findById(id).get();
 	}
